backend/src/Lawpal_Scraper/requirements.txt
requests
beautifulsoup4
faiss-cpu
sentence-transformers
spacy
numpy
